Introduce an original technique called `re-bucket fan-out`
In the previous commit, different index types were assinged to buckets for 8, 16 and 32 bits indices needs.

That they're needed 8, 16 or 32 bits for indices, depends if 8, 16 or 32 bits can index all the unique positions within a single object.

Here, with `re-bucket fan-out` a trade-off is attempted for each object.

For example, if a geometry would need 32-bits indices, the algorithm tries to downgrade the type of index to 16 bits (same for 16 to 8 bit indices).

By downgrading the index bucket type for a given geometry, the geometry is expanded (_fan-out_) to multiple sub-geometries, each using the downgraded indices bits.

A tradeoff appears then:

- as there is the need to _carry_ unique-positions from the geometry-to-downgrade to the fan-out geometries, the size of the positions array increases, we're duplicating unique positions
- but, at same time, the number of indices for the geometry remains constant now with half-space usage (as indices are downgraded from 32 to 16 bits, or 16 to 8 bits)

As there exists this fan-out positions expansion, and effectively the bucket type for the indices change, this technique is called `re-bucket fan-out`.

Rebucketting is not applied if it either:
- would increase the RAM usage for the geometry (this is calculated in the code)
- if it would fan-out, at the moment, to more than 8 sub-geometries. This relates to stability problems in case of some found massive re-buckettings.