Remove `TrianglesBatchingFlatColorRenderer` class.
This is based in the fact that the class was used to render geometry when the geometry had no normals available.
As the standard `TrianglesBatchingColorRenderer` is not using normals information (but calculating it in the Vertex Shader), it
automatically covers the use case of the removed class in this commit.